requirements.txt
openai
python-dotenv
azure-cosmos
rich
typer